+++ changelog.md
@@ -9,6 +9,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Unfinished Ideas
 
+## [0.1.3] - 2021-03-21
+
+### Added
+- Tests for NativeRegexSet
+- Tests for regexes that terminate due to end of string
+
 ## [0.1.2] - 2021-03-18
 
 ### Added